SO UOM Changes

+++ PAS_Database/dbo/Tables/Tables1/SalesOrderStockLine.sql
@@ -5,17 +5,17 @@
     [StockLIneId]         BIGINT          NOT NULL,
     [ItemMasterId]        BIGINT          NOT NULL,
     [ConditionId]         BIGINT          NOT NULL,
-    [Quantity]            INT             NULL,
-    [QtyReserved]         INT             NULL,
-    [QtyIssued]           INT             NULL,
+    [Quantity]            DECIMAL (18, 6) NULL,
+    [QtyReserved]         DECIMAL (18, 6) NULL,
+    [QtyIssued]           DECIMAL (18, 6) NULL,
     [AltPartMasterPartId] BIGINT          NULL,
     [EquPartMasterPartId] BIGINT          NULL,
     [IsAltPart]           BIT             NULL,
     [IsEquPart]           BIT             NULL,
-    [UnitCost]            DECIMAL (20, 2) NULL,
-    [ExtendedCost]        DECIMAL (20, 2) NULL,
-    [UnitPrice]           DECIMAL (20, 2) NULL,
-    [ExtendedPrice]       DECIMAL (20, 2) NULL,
+    [UnitCost]            DECIMAL (20, 6) NULL,
+    [ExtendedCost]        DECIMAL (20, 6) NULL,
+    [UnitPrice]           DECIMAL (20, 6) NULL,
+    [ExtendedPrice]       DECIMAL (20, 6) NULL,
     [MasterCompanyId]     INT             NOT NULL,
     [CreatedBy]           VARCHAR (256)   NOT NULL,
     [UpdatedBy]           VARCHAR (256)   NOT NULL,
@@ -30,6 +30,8 @@
     CONSTRAINT [FK_SalesOrderStockLine_SalesOrder] FOREIGN KEY ([SalesOrderId]) REFERENCES [dbo].[SalesOrder] ([SalesOrderId]),
     CONSTRAINT [FK_SalesOrderStockLine_StockLine] FOREIGN KEY ([StockLIneId]) REFERENCES [dbo].[Stockline] ([StockLineId])
 );
+
+
